// shared/types.d.ts
export type UserRole = 'agent' | 'admin' | 'customer';
export type AgentStatus = 'online' | 'away' | 'busy' | 'offline';
// ...
export interface User {
    id: string;
    email: string;
    name: string;
    role: UserRole;
    avatar?: string;
    status: AgentStatus;
    password?: string;
}